Flytt av äldre filer och införande av motionsmallen
Jag har strukturerat om lite och uppdaterat motionsmallen till fstildok
