+++ äldre/ekonomi/Autoverifikat/AutoVerifikatMall.tex
@@ -1,0 +1,102 @@
+
+%Här börjar dokumentet! All kommer ske automatiskt här nere!
+\documentclass[a4paper,10pt]{article}
+\usepackage[swedish]{babel}
+\usepackage{fstil}
+
+\title{Verifikation}
+\date{\vernr}
+\begin{document}
+
+
+\begin{Large}{\bf VERIFIKATION}\end{Large}
+\flushleft
+\vspace{0.5cm}
+
+\begin{tabular}{l p{1cm} p{5mm} l l}
+Verifikationsnr:& \large{\vernr{}} &&Verifikationsdatum:&\large{\verdate{}}\\
+\\ &&&Bokföringsdatum:& \large{\bokdate{}}\\
+\end{tabular}
+
+\vspace{2.5\baselineskip}
+
+Specifikation:
+\begin{tabular}{p{\textwidth}}
+\Specifikation
+\end{tabular}
+
+\begin{center}
+  \begin{tabular}{|p{4.5cm}|p{.8cm}|p{.4cm}|p{1.0cm}|p{1.6cm}|p{1.6cm}|}
+    \hline
+        \vspace{0pt}\large{Kontonamn} &\vspace{0pt}\small{Konto}& \vspace{0pt}\small{RS} & \vspace{0pt}\small{Projekt} & \vspace{0pt}\large{Debet} & \vspace{0pt}\large{Kredit} \\
+    \hline
+    \hline
+
+{\raisebox{-5pt}{\normalsize{\kontonamnA{}}}}
+&\centering {\raisebox{-6pt}{\large{\KontoA{}}}} 
+&\centering {\raisebox{-6pt}{\large{\RSA{}}}}
+&\centering {\raisebox{-6pt}{\large{\ProjektA{}}}}
+&\hfill     {\raisebox{-6pt}{\large{\DebetA{}}}} 
+&\hfill     {\raisebox{-6pt}{\large{\KreditA{}}}} 
+		\vspace{6pt}
+		\\
+    \hline
+{\raisebox{-5pt}{\normalsize{\kontonamnB{}}}}
+&\centering {\raisebox{-6pt}{\large{\KontoB{}}}} 
+&\centering {\raisebox{-6pt}{\large{\RSB{}}}}
+&\centering {\raisebox{-6pt}{\large{\ProjektB{}}}}
+&\hfill     {\raisebox{-6pt}{\large{\DebetB{}}}} 
+&\hfill     {\raisebox{-6pt}{\large{\KreditB{}}}} 
+		\vspace{6pt}
+		\\
+    \hline
+{\raisebox{-5pt}{\normalsize{\kontonamnC{}}}}
+&\centering {\raisebox{-6pt}{\large{\KontoC{}}}} 
+&\centering {\raisebox{-6pt}{\large{\RSC{}}}}
+&\centering {\raisebox{-6pt}{\large{\ProjektC{}}}}
+&\hfill     {\raisebox{-6pt}{\large{\DebetC{}}}} 
+&\hfill     {\raisebox{-6pt}{\large{\KreditC{}}}} 
+		\vspace{6pt}
+		\\
+    \hline
+{\raisebox{-5pt}{\normalsize{\kontonamnD{}}}}
+&\centering {\raisebox{-6pt}{\large{\KontoD{}}}} 
+&\centering {\raisebox{-6pt}{\large{\RSD{}}}}
+&\centering {\raisebox{-6pt}{\large{\ProjektD{}}}}
+&\hfill     {\raisebox{-6pt}{\large{\DebetD{}}}} 
+&\hfill     {\raisebox{-6pt}{\large{\KreditD{}}}} 
+		\vspace{6pt}
+		\\
+    \hline
+{\raisebox{-5pt}{\normalsize{\kontonamnE{}}}}
+&\centering {\raisebox{-6pt}{\large{\KontoE{}}}} 
+&\centering {\raisebox{-6pt}{\large{\RSE{}}}}
+&\centering {\raisebox{-6pt}{\large{\ProjektE{}}}}
+&\hfill     {\raisebox{-6pt}{\large{\DebetE{}}}} 
+&\hfill     {\raisebox{-6pt}{\large{\KreditE{}}}} 
+		\vspace{6pt}
+		\\
+    \hline
+        \multicolumn{4}{|r|}{\raisebox{-5pt}{TOTALT}} &\hfill{\raisebox{-6pt}{\large{\TotalDebet{}}}}
+							\vspace{.6\baselineskip} &\hfill{\raisebox{-6pt}{\large{\TotalKredit{}}}} \\
+    \hline  
+  \end{tabular}
+\end{center}
+
+
+\vspace{1cm}
+
+Kvitto/faktura finns på baksidan.
+\vspace{0.5cm}\\
+\begin{tabular}{p{6.2cm}p{6.2cm}}
+Sakattest & Ekonomisk attest\\
+\tabularnewline
+\tabularnewline
+\tabularnewline
+\tabularnewline
+\rule{5.2cm}{0.5pt} & \rule{5.2cm}{0.5pt} \\
+\sakattestNamn& \ekattestNamn\\
+\sakattestTitel& \ekattestTitel\\
+\end{tabular}
+
+\end{document}
