+++ äldre/ekonomi/skapa-verifikat/indata.txt
