+++ äldre/spexbiljetter/exempel/fysikalen.png
