+++ äldre/spexbiljetter/exempel/nobel2012/dieselverkstaden.txt
@@ -1,0 +1,295 @@
+#Format:
+%Sektion,Bänk,Plats,Sida
+Parkett,1,1,Höger
+Parkett,1,2,Höger
+Parkett,1,3,Höger
+Parkett,1,4,Höger
+Parkett,1,5,Höger
+Parkett,1,6,Höger
+Parkett,1,7,Höger
+Parkett,1,8,Höger
+Parkett,1,9,Höger
+Parkett,1,10,Höger
+Parkett,1,11,Vänster
+Parkett,1,12,Vänster
+Parkett,1,13,Vänster
+Parkett,1,14,Vänster
+Parkett,1,15,Vänster
+Parkett,1,16,Vänster
+Parkett,1,17,Vänster
+Parkett,1,18,Vänster
+Parkett,1,19,Vänster
+Parkett,1,20,Vänster
+Parkett,2,21,Höger
+Parkett,2,22,Höger
+Parkett,2,23,Höger
+Parkett,2,24,Höger
+Parkett,2,25,Höger
+Parkett,2,26,Höger
+Parkett,2,27,Höger
+Parkett,2,28,Höger
+Parkett,2,29,Höger
+Parkett,2,30,Höger
+Parkett,2,31,Vänster
+Parkett,2,32,Vänster
+Parkett,2,33,Vänster
+Parkett,2,34,Vänster
+Parkett,2,35,Vänster
+Parkett,2,36,Vänster
+Parkett,2,37,Vänster
+Parkett,2,38,Vänster
+Parkett,2,39,Vänster
+Parkett,2,40,Vänster
+Parkett,3,41,Höger
+Parkett,3,42,Höger
+Parkett,3,43,Höger
+Parkett,3,44,Höger
+Parkett,3,45,Höger
+Parkett,3,46,Höger
+Parkett,3,47,Höger
+Parkett,3,48,Höger
+Parkett,3,49,Höger
+Parkett,3,50,Höger
+Parkett,3,51,Vänster
+Parkett,3,52,Vänster
+Parkett,3,53,Vänster
+Parkett,3,54,Vänster
+Parkett,3,55,Vänster
+Parkett,3,56,Vänster
+Parkett,3,57,Vänster
+Parkett,3,58,Vänster
+Parkett,3,59,Vänster
+Parkett,3,60,Vänster
+Parkett,4,61,Höger
+Parkett,4,62,Höger
+Parkett,4,63,Höger
+Parkett,4,64,Höger
+Parkett,4,65,Höger
+Parkett,4,66,Höger
+Parkett,4,67,Höger
+Parkett,4,68,Höger
+Parkett,4,69,Höger
+Parkett,4,70,Höger
+Parkett,4,71,Vänster
+Parkett,4,72,Vänster
+Parkett,4,73,Vänster
+Parkett,4,74,Vänster
+Parkett,4,75,Vänster
+Parkett,4,76,Vänster
+Parkett,4,77,Vänster
+Parkett,4,78,Vänster
+Parkett,4,79,Vänster
+Parkett,4,80,Vänster
+Parkett,5,81,Höger
+Parkett,5,82,Höger
+Parkett,5,83,Höger
+Parkett,5,84,Höger
+Parkett,5,85,Höger
+Parkett,5,86,Höger
+Parkett,5,87,Höger
+Parkett,5,88,Höger
+Parkett,5,89,Höger
+Parkett,5,90,Höger
+Parkett,5,91,Vänster
+Parkett,5,92,Vänster
+Parkett,5,93,Vänster
+Parkett,5,94,Vänster
+Parkett,5,95,Vänster
+Parkett,5,96,Vänster
+Parkett,5,97,Vänster
+Parkett,5,98,Vänster
+Parkett,5,99,Vänster
+Parkett,5,100,Vänster
+Parkett,6,101,Höger
+Parkett,6,102,Höger
+Parkett,6,103,Höger
+Parkett,6,104,Höger
+Parkett,6,105,Höger
+Parkett,6,106,Höger
+Parkett,6,107,Höger
+Parkett,6,108,Höger
+Parkett,6,109,Höger
+Parkett,6,110,Höger
+Parkett,6,111,Vänster
+Parkett,6,112,Vänster
+Parkett,6,113,Vänster
+Parkett,6,114,Vänster
+Parkett,6,115,Vänster
+Parkett,6,116,Vänster
+Parkett,6,117,Vänster
+Parkett,6,118,Vänster
+Parkett,6,119,Vänster
+Parkett,6,120,Vänster
+Parkett,7,121,Höger
+Parkett,7,122,Höger
+Parkett,7,123,Höger
+Parkett,7,124,Höger
+Parkett,7,125,Höger
+Parkett,7,126,Höger
+Parkett,7,127,Höger
+Parkett,7,128,Höger
+Parkett,7,129,Höger
+Parkett,7,130,Höger
+Parkett,7,131,Vänster
+Parkett,7,132,Vänster
+Parkett,7,133,Vänster
+Parkett,7,134,Vänster
+Parkett,7,135,Vänster
+Parkett,7,136,Vänster
+Parkett,7,137,Vänster
+Parkett,7,138,Vänster
+Parkett,7,139,Vänster
+Parkett,7,140,Vänster
+Parkett,8,141,Höger
+Parkett,8,142,Höger
+Parkett,8,143,Höger
+Parkett,8,144,Höger
+Parkett,8,145,Höger
+Parkett,8,146,Höger
+Parkett,8,147,Höger
+Parkett,8,148,Höger
+Parkett,8,149,Höger
+Parkett,8,150,Höger
+Parkett,8,151,Vänster
+Parkett,8,152,Vänster
+Parkett,8,153,Vänster
+Parkett,8,154,Vänster
+Parkett,8,155,Vänster
+Parkett,8,156,Vänster
+Parkett,8,157,Vänster
+Parkett,8,158,Vänster
+Parkett,8,159,Vänster
+Parkett,8,160,Vänster
+Parkett,9,161,Höger
+Parkett,9,162,Höger
+Parkett,9,163,Höger
+Parkett,9,164,Höger
+Parkett,9,165,Höger
+Parkett,9,166,Höger
+Parkett,9,167,Höger
+Parkett,9,168,Höger
+Parkett,9,169,Höger
+Parkett,9,170,Höger
+Parkett,9,171,Vänster
+Parkett,9,172,Vänster
+Parkett,9,173,Vänster
+Parkett,9,174,Vänster
+Parkett,9,175,Vänster
+Parkett,9,176,Vänster
+Parkett,9,177,Vänster
+Parkett,9,178,Vänster
+Parkett,9,179,Vänster
+Parkett,9,180,Vänster
+Parkett,10,181,Höger
+Parkett,10,182,Höger
+Parkett,10,183,Höger
+Parkett,10,184,Höger
+Parkett,10,185,Höger
+Parkett,10,186,Höger
+Parkett,10,187,Höger
+Parkett,10,188,Höger
+Parkett,10,189,Höger
+Parkett,10,190,Höger
+Parkett,10,191,Vänster
+Parkett,10,192,Vänster
+Parkett,10,193,Vänster
+Parkett,10,194,Vänster
+Parkett,10,195,Vänster
+Parkett,10,196,Vänster
+Parkett,10,197,Vänster
+Parkett,10,198,Vänster
+Parkett,10,199,Vänster
+Parkett,10,200,Vänster
+Parkett,11,201,Höger
+Parkett,11,202,Höger
+Parkett,11,203,Höger
+Parkett,11,204,Höger
+Parkett,11,205,Höger
+Parkett,11,206,Höger
+Parkett,11,207,Höger
+Parkett,11,208,Höger
+Parkett,11,209,Höger
+Parkett,11,210,Höger
+Parkett,11,211,Vänster
+Parkett,11,212,Vänster
+Parkett,11,213,Vänster
+Parkett,11,214,Vänster
+Parkett,11,215,Vänster
+Parkett,11,216,Vänster
+Parkett,11,217,Vänster
+Parkett,11,218,Vänster
+Parkett,11,219,Vänster
+Parkett,11,220,Vänster
+Parkett,12,221,Höger
+Parkett,12,222,Höger
+Parkett,12,223,Höger
+Parkett,12,224,Höger
+Parkett,12,225,Höger
+Parkett,12,226,Höger
+Parkett,12,227,Höger
+Parkett,12,228,Höger
+Parkett,12,229,Höger
+Parkett,12,230,Höger
+Parkett,12,231,Vänster
+Parkett,12,232,Vänster
+Parkett,12,233,Vänster
+Parkett,12,234,Vänster
+Parkett,12,235,Vänster
+Parkett,12,236,Vänster
+Parkett,12,237,Vänster
+Parkett,12,238,Vänster
+Parkett,12,239,Vänster
+Parkett,12,240,Vänster
+Parkett,13,241,Höger
+Parkett,13,242,Höger
+Parkett,13,243,Höger
+Parkett,13,244,Höger
+Parkett,13,245,Höger
+Parkett,13,246,Höger
+Parkett,13,247,Höger
+Parkett,13,248,Höger
+Parkett,13,249,Höger
+Parkett,13,250,Vänster
+Parkett,13,251,Vänster
+Parkett,13,252,Vänster
+Parkett,13,253,Vänster
+Parkett,13,254,Vänster
+Parkett,13,255,Vänster
+Parkett,13,256,Vänster
+Parkett,13,257,Vänster
+Parkett,14,258,Höger
+Parkett,14,259,Höger
+Parkett,14,260,Höger
+Parkett,14,261,Höger
+Parkett,14,262,Höger
+Parkett,14,263,Höger
+Parkett,14,264,Höger
+Parkett,14,265,Höger
+Parkett,14,266,Höger
+Parkett,14,267,Höger
+Parkett,14,268,Vänster
+Parkett,14,269,Vänster
+Parkett,14,270,Vänster
+Parkett,14,271,Vänster
+Parkett,14,272,Vänster
+Parkett,14,273,Vänster
+Parkett,14,274,Vänster
+Parkett,14,275,Vänster
+Parkett,14,276,Vänster
+Parkett,15,277,Höger
+Parkett,15,278,Höger
+Parkett,15,279,Höger
+Parkett,15,280,Höger
+Parkett,15,281,Höger
+Parkett,15,282,Höger
+Parkett,15,283,Höger
+Parkett,15,284,Höger
+Parkett,15,285,Höger
+Parkett,15,286,Vänster
+Parkett,15,287,Vänster
+Parkett,15,288,Vänster
+Parkett,15,289,Vänster
+Parkett,15,290,Vänster
+Parkett,15,291,Vänster
+Parkett,15,292,Vänster
+Parkett,15,293,Vänster
