+++ äldre/spexbiljetter/exempel/nobel2012/nobel-logo-bw.png
